fix reading acc and gyro values for 32-bit

+++ applications/imu/mpu6050/mpu6050.cpp
@@ -356,9 +356,10 @@ Vector MPU6050::readRawAccel(void)
     uint8_t zha = buffer[4];
     uint8_t zla = buffer[5];
 
-    ra.XAxis = xha << 8 | xla;
-    ra.YAxis = yha << 8 | yla;
-    ra.ZAxis = zha << 8 | zla;
+    ra.XAxis = (int16_t)(xha << 8 | xla);
+    ra.YAxis = (int16_t)(yha << 8 | yla);
+    ra.ZAxis = (int16_t)(zha << 8 | zla);
+
 
     return ra;
 
@@ -434,9 +435,9 @@ Vector MPU6050::readRawGyro(void)
     uint8_t zha = buffer[4];
     uint8_t zla = buffer[5];
 
-    rg.XAxis = xha << 8 | xla;
-    rg.YAxis = yha << 8 | yla;
-    rg.ZAxis = zha << 8 | zla;
+    rg.XAxis = (int16_t)(xha << 8 | xla);
+    rg.YAxis = (int16_t)(yha << 8 | yla);
+    rg.ZAxis = (int16_t)(zha << 8 | zla);
 
     return rg;
